[BUGFIX] Update most recent workspace record version
This patches the new hooks setup to process the most
recent version of a record, when moving any record.

This extra update is necessary for the versioned record
to have the correct relationship values, which in turn
makes sure that once published, the published record
will have the correct relationship.

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/