[BUGFIX] Fix FlexForm empty nodes removal
This commit changes the way Flux cleans a FlexForm. One of the things done by
this cleaning function was to delete all the empty `sheet` elements, because it
makes TYPO3 throw an exception in some cases.

But there was a problem in the way this feature was handled: looping on the
elements and removing elements at the same time could cause nodes to be skipped.
For instance, if three `sheet` nodes were empty in a row, the second one was
skipped because the first one was removed and the looping pointer was not
correctly updated.

The implemented solution to this (big) issue is to first loop on all `sheet`
nodes, and store the one which _should_ be deleted first, _then_ loop again on
this list and remove the nodes one by one.