[BUGFIX] Split TCA post processing hook into two (#1293)
The extTablesInclusion-PostProcessing hook behaves
unpredictably in FE when no BE user is logged in and
caches expire. In order to trigger the hook consistently,
the same hook subscriber must be registered in two
places, depending on BE or not BE context.

Specifically, prevents missing TCA in the case described
above - and allows template-as-CType registrations to
function correctly.