[FEATURE] New design for collapsing columns
Displaying the toggler in the header of the element.
Right next to the edit / hide / delete / buttons.

Bugfix in fluxCollapse.js
Array.isArray() is used to check whether the cookie variable is of type "array" and to avoid javascript errors.
Optimizing Cookie handling; remove "null" items from cookie array.

Tested in TYPO3 7.6.15, 8.6.0-dev