[BUGFIX] Prevent moving records into any child of self
Fixes a potential issue where users may attempt to
cut a record and paste it after another record that
is already a child of the record being moved, at any
depth. Allowing this to happen results in an infinite
recursion when attempting to render the result.

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/