[BUGFIX] Refactor template-as-CType TCA to allow overrides
The patch changes the logic used to register Flux templates
as new content types. The problem with the existing logic
is that it does not allow TCA customisations unless those
are also done with the "table configuration post processing"
hook that Flux uses.

Moving the TCA creation - which does not depend on the
context that the plugin registrations depend on - allows the
TCA that gets generated as boilerplate for such CTypes,
to be overridden in ext_tables / Configuration/TCA/Overrides.