[BUGFIX] Replace exceptions with fallbacks
When registering templates as CType the current logic
requires you to have a controller class associated with
the content type/template. However, the ContentController
from Flux is perfectly capable of rendering the elements
because they use a well-configured Provider.

Exceptions are replaced with dev-logging and a fallback
to use Flux's ContentController if the custom controller
could not be detected by class name.