[BUGFIX] Do not cache TS if all TS is empty
At certain points in the runtime it is possible to use
FluxService for various operations which normally
require TS to be present. In such case Flux assumes
that the calling code does not depend on any TS.

However, the same operation may at other points
in the runtime still check for and use TS if it is present.

In this case, the first and empty TS array would have
been cached. With this patch, such empty TS arrays
are no longer cached and TS gets re-checked until
it becomes available - then it gets cached.

In addition, the method which returns view configuration
for a specific extension key, now checks if the full TS
array is empty and if it is, it too avoids caching the
configuration and instead returns the default configuration.