[BUGFIX] Move icon rendering functions to TCA-defined userFunc
This allows the icons which get resolved from template's
settings or by naming convention, to be generated as a
true custom icon identifier, instead of the previous method
of CSS-overlaying icon on top of default icon.

Solves an issue with TYPO3 master and 8.6 where the
CSS method caused the icon to not be easily clickable
(as in, hidden below an invisible yet clickable element).

This also means we get rid of some fluidcontent-only
styles which did not belong in Flux.