[BUGFIX] Prevent large payloads in XHR url
Prevents passing the entire record or temporary record
as part of the DS identifier. Passing too much data in
this identifier can cause the XHR request to fail. Instead,
we pass either just the record UID (and thaw the record
later on) or we pass only the fields which are defined in
the TCA control useColumnsForDefaultValues. We now
emulate the amount of data you'd receive in the new
content wizard (e.g. only specially marked columns).