[BUGFIX] Add `tt_content` in list of tables affected by `new` command (#1316)
This commit fixes a severe issue that would return a completely wrong page id
for a given record upon creation.

This could lead to really severe issue. For instance the wrong TypoScript tree
was loaded for a new content using FluidContent, only on creation.

When creating a new `tt_content`, the content should NOT be fetched to get its
pid. Seems logical since the content is not created yet, and is not in the
database yet.