[BUGFIX] Resolve move placeholder as relative-to record
When performing relative-to-record cut and paste operations
in a workspace, the new positioning information would be
taken from the original record if the target relative-to record
was a placeholder.

Resolving the potential placeholder before reading the new
positioning values solves the issue.

In addition, two checks are added which prevents problems
on TYPO3 LTS because the input records differ from
TYPO3v8. On TYPO3v8 the placeholder resolving is not
necessary as the input record is already the placeholder
there. The potential problems are prevented wen finding
the new positioning values and when determining if the
record being moved would end up becoming a child of
itself (TYPO3 LTS yielded a false positive due to the
mentioned differences in input).

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/