[BUGFIX] Force workspace publishing to start with children
This patch prevents an issue where moving child records
outside of a parent and then deleting the parent while in
a workspace caused the children to be deleted as well.

This happened because parent records are intentionally
processed first when a workspace is published.

The solution is to re-sort the command map array before
the DataHandler starts the sequence of "swap" version
commands, so the list is sorted with children first.

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/