[TASK] Update DragDrop.js
Using the new DragDrop.js from TYPO3 8, which have changed from the jQuery UI sortable to droppable.
Copying the TYPO3 8 css for drag and drop.
Better drag and drop feeling.
Worked also in TYPO3 7.6 ;-)

The copy feature during drag and drop (holding the cmd-key) in TYPO3 8 is disabled in TYPO3 7.6.
This is done with a check if the localized text for 'dragdrop.copy.message' exists.

Fixed a z-index bug during dragging an element outside from a grid column.
Hide the dragged element in the column. This is useful for an element which contains a flux grid.