[BUGFIX] Fix saving of child records in workspace
A regression was added with the patch which sorts
command map operations to put child record operations
before parent record, to solve the issue of publishing
a deletion of a parent after moving children outside the
parent element. The approach chosen to merge arrays
of command operations removed the keys - and the keys
are needed by the internal logic. Incorrect keys triggered
an error which seemed completely unrelated:

`Could not be edited in offline workspace in the branch where found (...)`

By preserving the keys this error is prevented and saving
of child record works again.

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/