[BUGFIX] Force sorting from versioned record to placeholder
This patch works around side effects caused by the use
of IRRE relation in TCA to define relationship between
parent and child records.

Briefly described, the problem is that when you move a
record in a workspace, the sorting value of child records
in the LIVE workspace get updated with a sequential
sorting number (even though editing happened in a
workspace, LIVE record is updated). Furthermore, the
move placeholders for child records would sort in reverse
with new sorting values (normal sorting value spacing, but
values all reversed compared to original order).

The patch fixes the second of these symptoms by forcing
the sorting value of the placeholder to be the same as the
one set in the versioned record - which still contains the
proper, original sorting value. The result is that live records
do still get updated, but sorting order is preserved albeit
incorrect (sequential, not spaced), and when workspace is
published, the sorting value gets set to whichever sorting
value the move placeholder has.

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/