[BUGFIX] Move child placeholders when parent moves
Due to a bug in TYPO3, cascading operations on IRRE
such as the one Flux uses, does not cascade updates
to the child records when they are move placeholders.

The symptom is that moving a parent in a non-live
workspace from one page to another does not update
the child content when the child content is a placeholder.

The patch solves this by manually and recursively
dispatching SQL updates to fix the incorrect PID value
of such placeholders.

Fix kindly sponsored by MSI Chicago - http://www.msichicago.org/