[BUGFIX] Allow any TCA field to be mirrored in Flux form
Using the `$tableName.$fieldName` format as name
of a Flux form field causes the value to be saved to that
field. The check before would ignore the field if the
original field was not also posted - after this patch the
check only cares if the field is defined in TCA.

Note that attempting to update a field which a user has
no access to causes an error.

Close: #1250