[FEATURE] Support arbitrary single-argument function transformations
Entering a function reference in “transform” properties of fields is now possible:

    transform=“TYPO3\CMS\Core\Resource\FileRepository->findFileReferenceByUid”

For example, transforms a single UID of a file reference to an actual FileReference object.
Supports core and third party functions alike, but the function must require no more than a single mandatory argument (and can have as many optional ones after that as it wants).

Please note that this only works when the input value is a UID or list of UIDs (and the function accepts lists as input). Functions which for example require an array of relations must still be processed manually - in the case of IRRE or M:N relations the local field contains a count of relations, not a UID.

Close: #1425