[BUGFIX] Merge with instead of replace existing DS
Since refactoring of the FlexForm internals in FormEngine, the hook that manipulates DS works slightly differently. A key difference is that the original DS is no longer available to the hook method which *parses* the DS identifier.

The solution is to generate the original identifier and pass that as part of the Flux identifier, then use the existing FlexFormTools logic to parse that identifier as starting point for the DS array that is then passed to Flux’s providers.

Close: #1409