[BUGFIX] Fix cache key and return type for data from templates
Before this patch, the cache key would always pass "_all" as the name of the extracted variable - after the patch, the actual name passed to the function is used. The return type is also fixed to be NULL if variable does not exist.

Logic is still prepared for the eventual ability to return all variables from a ViewHelperVariableContainer instead of always having to return a specific variable ("_all" used as virtual variable name in cache key when caching all variables).