[FEATURE] Add vast collection of hooks
This patch adds 44 hooks to pretty much every part of Flux's internal logic, allowing you to subscribe to hooks and, for most hooks, manipulate return data.

Hook subscribers must implement `\FluidTYPO3\Flux\Hooks\HookSubscriberInterface` which contains a single method that receives the hook name and a payload of data, and must return the data with or without modifications.

There are two ways to register hook subscribers:

* `\FluidTYPO3\Flux\Hooks\HookHandler::subscribe()`
* `$GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['flux']['hooks'][$hook][$subscriberClass] = $subscriberClass;`

The former is useful if you need a strict validation of input arguments. The latter is useful if you need the hooks to be configured in, for example, AdditionalConfiguration.php.

See https://github.com/FluidTYPO3/flux/tree/development/Classes/Hooks/HookHandler.php for a list of hooks.