[FEATURE] Use colPos instead of tx_flux_column
This is a breaking change!

With this patch, Flux switches from using tx_flux_column
and tx_flux_parent to determine relations between parent
and child records, to using a pure colPos based solution.

The new behavior is designed in the following way:

* Every flux:grid.columm *must* now use “colPos” argument
* The “colPos” should be treated exactly like in page templates
  so you use numbers from 0-99 (99 is the new maximum!)
* The final colPos value is then calculated based on a math
  operation that includes the parent’s UID.

This means that a flux:grid.column with a colPos of 4,
when you use it in a grid inside a content record with
an UID of 123, the real colPos value becomes 12304.

This value can then be reverse calculated to return
the UID of the parent and the colPos in the template.

In order to render columns, flux:content.render and
the .get variant have a new “colPos” argument which
can be used with the virtual colPos value, for example
“4”, to render the real column based on the UID of the
current parent record.

It is also possible to use math to calculate a colPos
value and use this with v:content.get and v:content.render
if this is preferred. It will then of course be necessary
to manually keep track of when to use a “raw” colPos
value (in page templates) or a calculated colPos value
(in content templates).

Note that the “list nested content” feature is also removed
- list view will show all content without possibility of
messing up sorting, which was the reason for having
child content hidden in list view in the first place.