[BUGFIX] Make edit icon on child content edit single record
The PageLayoutView instance was misconfigured; when clicking the “edit” icon of a content record, the action that was performed was actually to edit that record plus the next two records based on sorting and colPos value.

This patch fixes that so a single record is edited, like it happens in root columns.

Close: #1586