[BUGFIX] Avoid recursive copy if any parent did not copy
Prevents dispatching another recursive call to copy
child records, if for some reason a record could not
be copied. The exact reason is unknown but will be
logged by the DataHandler - so we log an additional
message that the failure to copy also meant no child
records could be copied.

References: #1589