[BUGFIX] Handle expanded records from FormEngine
This patch adds an additional check for a UID value possibly being contained in an array, which happens when a record is passed from FormEngine to the Flux override for BackendLayoutView.

Though this doesn’t happen in core code, it may occur in third party code that re-uses the record FormEngine expanded and calls core functions like preview rendering with that record as input. Flux can easily handle the case.

References: #1645