[FEATURE] Allow translating child content with localize button
Adds an override for the LocalizationController which
automatically adds an ad-hoc column entry for
records which are not located directly in a page column
and which have a colPos value above the 100 limit.

The result is that child content records will be shown
with the ability to sync each individual child column,
without including unintended content from page columns.

Close: #1662