[BUGFIX] Allow null returns when loading DB record
The BackendLayoutView override class assumes that
records are possible to read from DB because it is
almost only used in the PageLayoutView which can only
show records that exist.

However, the class is also consulted for colPos value
lists when creating tt_content records as IRRE so it must
be allowed to silently ignore possibly missing parent
or child records for parents that have not been saved.

Close: #1598