[FEATURE] Add TCA displayCond user function to hide "flex" field
Adds a user function that can be defined as displayCond
for a TCA field of type "flex" and will hide the entire field
if there are no fields in the DS resolved by Flux.

Example:

```
$GLOBALS['TCA']['pages']['columns']['tx_fed_page_flexform']['displayCond'] =
    'USER:' . \FluidTYPO3\Flux\Integration\FormEngine\UserFunctions::class
    . '->fluxFormFieldDisplayCondition:pages:tx_fed_page_flexform';
```

Note the last two segments of the function reference must be
the table name and the field name (since these are unfortunately
not passed to the user function automatically by TYPO3).