[BUGFIX] Fixed language handling for nesting (#1716)
Having a parent record with sys_language_uid -1 and a child
record with sys_language_uid 0 will result in the child record
not being shown. This happens because till now the respected
sys_language_uid of the child record will be set to the parent
record, which would be -1. This results in a query looking for
the record with sys_language_uid = -1, which dosn't exist.

The normal behaviour should respect the currently selected
language in the Backend as in the Backend/PageLayoutController
shown in Line 926 (TYPO3 8).