[REMOVAL] Deprecate Field ViewHelper API
Deprecates the ViewHelpers used to generate fields in
forms, and schedules them for removal in Flux 10.

The new method for generating fields is:

    <flux:field type="select" name="myfield" label="My Label"
        config="{foreign_table: 'pages'}" />

And using PHP:

    $field = Field::create(
        [
            'type' => 'select',
            'name' => 'myfield',
            'label' => 'My Field',
            'config' => [
                'foreign_table' => 'pages',
            ]
       ]
    );

The "config" array then supports every TCA option
for the selected field type and means Flux no longer
has to be updated when TYPO3 changes the way
TCA is processed/validated.

Close: #1465
Close: #1734
Close: #1730
Close: #1701
Close: #1641
Close: #1547