[BUGFIX] Disable drop-in templates for unexpected public paths
Avoids creation of drop-in template types when the
public path of TYPO3 is set to an unexpected value.

The ignore happens silently since there may be custom
reasons to set public path to a value that's not inside
the projet root, or is not a relative path - such a setup
may work with custom sites but is not compatible with
Flux's drop-in template logic.

Close: #1736