[BUGFIX] Do not update originals when copying children
Due to a misunderstanding of how DataHandler processes an
"update" property on a copy/copyToLanguage command, the
original versions of copied child records would be updated
when the intention was to update the copied versions of those
records.

In order to solve this, the adjustments to copies of records had
to be moved to an execution point after all database records
had been processed, and detection of the right parent record
had to be done based on what occurred during the copy action
(copy of parent, copy of individual child, and copyToLanguage
of both of those cases).

Close: #1764