[BUGFIX] Ensure numeric and sequential keyed array
Ensures that the columnList array that is handed off
to the LocalizatonController is built with numeric
indexes and consistently sequential keys. If the array
is incorrectly formatted and for example has holes
in numeric keys, it will be encoded as a JSON object
instead of a JSON array, causing the JS that handles
the localization dialog to fail with an "method forEach
does not exist".