[BUGFIX] Adjust colPos for recursively localized child content
Copying a page with a container that has child content which
has translations fails to update the colPos value of the copy
of the translated child content. The colPos of the translated
child element still refers to the original version of the copied
element.

By regenerating the colPos value based on the colPos value
of the original language version of the copied element this
problem is resolved.

Close: #1819