[FEATURE] Revive Outlet feature
Revives the Outlet feature by cleaning up the implementation
and removing the form fields integration so that Outlet/Pipe
is now exclusively configured at Form level. Adds TYPO3 10.4
support to the EmailPipe, and adds a "property" argument for
the TypeConverterPipe to allow converting only a single
property in the posted form data array.

Close: #1379
Close: #1095