[FEATURE] Upgrade wizard for problematic content element "sorting" values
Copying pages will fail to adjust the colPos values when the "sorting" value
of child elements is larger than that of its parent record.
This leads to "lost" content records that are visible in the list view,
but not in the web view in the TYPO3 backend.

This problem happens at least when migrating from TYPO3 v7 to v8 while
upgrading to flux 9.5.0.
Old content records exhibit this problem, while newly created content
is fine - TYPO3 takes care of modifying the "sorting" values correctly.

This wizard fixes the problematic sorting values by finding the largest
child sorting value and adjusting the parent's "sorting" accordingly,
as well as all following content element siblings.

Resolves: https://github.com/FluidTYPO3/flux/issues/1824