[BUGFIX] Sanitise Form instance sorting behaviour
Sorting, e.g. in new content wizard, is no longer exclusively based on the "sorting" option.

Instead, one of the following is used:

* Sorting option value, if set.
* Otherwise, Form's associated template filename (if Form is template-file-based).
* Otherwise, Form's "id" attribute.

Note that any Form which specifies a sorting value (which is an integer) will naturally
be sorted to top because integers sort before alphabet characters which may be used
in template file name or Form id. This is true regardless of how large the integer
sorting value is.

Close: #2057