[BUGFIX] Remove workaround for incompletely booted CLI contexts
Turns out, the problem that was worked around is exclusively caused
by  the TYPO3 Console package in versions below 8, since these versions
contain a special CLI command "typo3cms" which does not correctly
dispatch the same events that the core's "typo3" command will dispatch.

Since the workaround caused *separate* issues on TYPO3v12, the
workaround is removed.

Should you run into a problem similar to the one described in the
doc comment that is now removed, the solution is:

* Do not call "typo3cms" in any context that will trigger a nested
  call to TemplateService. Call "typo3" instead!
* Whenever possible, raise the version of TYPO3 Console to at
  least version 8 since this properly wraps it in a command runner
  that will dispatch the events which Flux depends on.