[BUGFIX] Prevent warnings if DataHandler received record has no language
Warnings may trigger if the current BE user does not have access
to the language column of a record being moved. When this happens,
Flux will now attempt to read the language UID from the record as
stored in DB, or will default to the default language UID of zero.