[FEATURE] Add inheritance mode control
Adds an extension configuration option and corresponding Form option
and option ViewHelper which allows a site to control how Flux handles
inheritance of FlexForm values (specific to pages).

Normally, Flux will only allow FlexForm values to be inherited from a
parent page to children if the child uses the same page layout (template)
as the parent page(s). This new option allows changing this inheritance
mode from the default "restricted" to an "unrestricted" mode which
allows FlexForm values to be inherited even if the child and parent uses
different page layouts.

This is useful when you have different page templates which use the
same FlexForm fields (for example, a shared set of fields) and you wish
to inherit such fields from parents even if you use different page
layouts on the child/parent(s).

There is a similar Form option which allows setting this inheritance
mode on a per-template basis. Use the new ViewHelper
flux:form.option.inheritanceMode for this, with either "restricted"
or "unrestricted" as value. If no mode is specified on a template then
the inheritance mode defined in extension configuration will be used.