[BUGFIX] Jump through ridiculous hoops imposed by inconsistent TYPO3 behavior
DataHandler will call two entirely different sets of methods when copying
a page vs. copying a different record. One impact of this difference is that
the record used for resolving a FlexForm DS will not contain a UID when
copying a page. Combined with how page templates are inherited in Flux
this results in a strange behavior where DataHandler would copy any
relations (files etc.) defined in the FlexForm but would not remap those
copied records to belong to the copy of the original page. The source
of that problem is that Flux is unable to return a DS because of the
incomplete input record.

The above only happens when the page record does not itself select
a page layout to use for the specific page. It only happens for inherited
page layouts.

Because of this, Flux has to introduce a special case when building the
FlexForm DS, to check if we are working with a page record. If we are,
we read the page layout configuration from the rootline using the "pid"
of the input record.

Close: #2146