[FEATURE] Add custom page layout selector
Note: Not available on TYPO3v10.

Adds a replacement page layout selector which can be
enabled instead of the default TCA "select" type.

The page layout selector is enabled in extension
configuration and can be configured through TCA
overrides (in Configuration/TCA/Overrides/pages.php
of an extension that depends on Flux):

```
// For both tx_fed_page_controller_action and tx_fed_page_controller_action_sub
// and can be configured individually, e.g. to make sub-layout icons smaller
// or only render titles/descriptions for "this page" layouts.

$GLOBALS['TCA']['pages']['columns']['tx_fed_page_controller_action']['config']['iconHeight'] = 200;
$GLOBALS['TCA']['pages']['columns']['tx_fed_page_controller_action']['config']['titles'] = true;
$GLOBALS['TCA']['pages']['columns']['tx_fed_page_controller_action']['config']['descriptions'] = true;
```