[BUGFIX] Fix localization record merging logic.
Update the logic to correctly merge localized records by checking the presence of a '_LOCALIZED_UID' in the content object's data. This change ensures that records are only merged when a valid localized version is available.

I extracted $this->recoredService->getSingle(...) from the array_merge function because it had never merged the localised record on my system here.
Therefore, getSingle is saved in an extra variable and then merged if it exists.