[BUGFIX] Correct various lookup attempts to find current record
Fixes mainly a problem with Flux in an INT uncached context where
the resolved request and therefore content object data was not the
expected one (returned page row from inside a content context).

This fixes that problem by more consistently handing over the request
to the lookup methods and using the content object data when it is
availble in the content object.